+++ R/utils.R
@@ -375,7 +375,7 @@ XtX_pseudoinv_sqrt <- function(mat, mat.is.XtX = FALSE, tol = .Machine$double.ep
   
   if (nrow(mat) == 0 && ncol(mat) == 0)
   {
-    warning(paste("Cannot calculate pseudoinverse:",
+    message(paste("Degenerate null covariance (rank 0 Gram matrix):",
       "perhaps all covariates are constant (within strata)?",
     collapse="\\n"))
     ans <- matrix(NA_real_, 0, 0)